fix(frontend): rename IconNetwork to IconNetworks for pluralization consistency
Matches the folder's existing convention of pluralizing multi-item
icons (e.g. IconLayers).
