+++ src/frontend/src/lib/components/icons/lucide/IconNetworks.svelte
