+++ src/frontend/src/lib/components/liquidium/LiquidiumInfoBox.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import IconLineChart from '$lib/components/icons/lucide/IconLineChart.svelte';
-	import IconNetwork from '$lib/components/icons/lucide/IconNetwork.svelte';
+	import IconNetworks from '$lib/components/icons/lucide/IconNetworks.svelte';
 	import IconShieldCheck from '$lib/components/icons/lucide/IconShieldCheck.svelte';
 	import StakeContentSection from '$lib/components/stake/StakeContentSection.svelte';
 	import ExternalLink from '$lib/components/ui/ExternalLink.svelte';
@@ -88,7 +88,7 @@
 			<FactBox>
 				{#snippet icon()}
 					<span class="rounded-full bg-warning-subtle-20 p-4 text-primary">
-						<IconNetwork size="24" />
+						<IconNetworks size="24" />
 					</span>
 				{/snippet}
